[celery] Renamed celery.py to task.py for consistency
Consistency with celery best practices
