+++ openwisp_firmware_upgrader/tasks.py
